Fix Travis build
Install PyYaml to fix use of the real make-zip script in the tests.
Instally Pylnit via PIP to get a more recent version.
Print the Pylint version before running.
Use print as a function since the most recent version of pylint now
appears to error about this (but we don't expect it to).

Change-Id: I975c1929bd7bc014343b91c2499d395569a8e661