Log an error for missing lib_robot revisions
When the lib_robot revision for a team is misconfigured we want to
know about this. Clearly using the main state of the repo is a valid
fallback as we still want to emit a usable robot.zip at the team,
so just log an error for us to see.

Change-Id: I33547d11af676ae6cc995cd6428eae4885b7f705