Increase the Pylint timeout to 5s
It seems that for larger files pylint takes a while to run even
when it's operating normally. On saffron a 1400 line file regularly
hits the previous 2s limit, and takes about 1.2s on my machine.

Five seconds therefore ought to allow it some more leeway while still
ensuring we respond to the client reasonably promptly.

Change-Id: I795c3524dfc90fd59acf9c21e68b9eaccf187522