Fix #3002: Make IDE polling more performant
Add an (optional) parameter which limits the returned revision
information to only a single project. Use this, along with better
poll setup structuring, in the front-end so that the overall load
on the server is reduced.

Local testing for a team with 30 projects showed that this reduced
the poll/poll response time from ~120ms to <10ms.

Change-Id: I7ff48dbdac9a1322a76e620f6997c1ee61062dbb