Ensure that poll results don't trigger pointless updates
This fixes a long-standing bug where the response to the first
poll would always trigger the project log and file list to be
reloaded.

While it's correct that the first poll response signals that
something has changed (since it has no way to know if the first
batch of data is really different to what came before), there was
nothing that was able to work out that there hadn't really been a
change. This commit adds the project revision to the proj/info
endpoint so that we can compare this to the value from the polls.
This does make the proj/info endpoint someswhat slower (3ms becomes
8ms or so), but since we save two redundant calls I think this is
worth it.

Also fix a bug where the current project revision wasn't being reset
when the project changed. It turns out that newly created projects
can have the same revision if created close enough together.

Change-Id: I7be7ba00047143dab36faa2bd546f9c1e05600d3