Don't keep the file extension of uploaded files
Having remove the code which looked like it needed this (but which
wasn't actually used) we can drop this logic and thus simplify our
handling of the uploaded files. PHP's 'getimagesize' seems fine
getting the MIME from files without extensions, so don't even bother
providing an extension for the temporary copy of the image.

This also means that we don't have to worry at all about the validity
of the file name the user provided since it's never used.

Change-Id: Ib82fdb2423736fd2817cd0fc47019dc9b58eb3de