Change to throwing on bad project names
The previous solution of returning FALSE wasn't really working since
the only source-code usage of the method didn't check it. Throwing
an exception better matches up to what happens if there are other
forms of bad name requested and since this is now handled by the
GitRepository class it fits better there too.

As a side-effect, we now also get a much better error message.

Change-Id: I5e1cd6cf21c30747300f1838e4ebc7cfac4bfddc