Add support for using pylint from a Virtualenv
Anoyingly, pylint doesn't version its dependencies so we're instead
forced to do this for it by requiring specific versions of the
packages it uses. Add a requirements file for this and install that
into the virtualenv we create.

Travis seems to have issues installing the latest version of pylint
into the main system, so hopefully this will provide some insulation.

Change-Id: I2d1a3541f5e84de9cc7803060cb46957ffc1fde2