Fix #2103: Add team healthcheck script
This script has been around for a while, but not anywhere useful.
Add a new scripts folder and add it so that it's available to all.

The script itself iterates over the master repos for each team,
looking at the time of the last commit to each. The most recent
is chosen as a reference and then the values for each team are
output. The output can be sorted by either TLA or activity date.

Change-Id: I6cb26bf5da9789d2da730615f3f2d3fe90cbd6ff