Fix #2174: Warn users if exporting projects with local changes
This doesn't account for files which are closed but have autosaved
changes, though these seem to be less likely to cause an issue.
Having detected this issue we focus on one of the files which has
such changes so that the user can look at what the changes are,
and/or save the file.

Change-Id: Id4ee65e7347daf4725ead208d7c5fb22120179f3