Have search also look in open buffers
This removes its reliance on the buffers having been autosaved and
should also make the results more correct. Weird things could happen
if the user has an older version of a file open, but that would
already have given a wrong result -- it's now just wrong in a
different way.

Change-Id: I3ad258d7ef187b7d9fd47afd4e9166a171c8672b