Avoid redundant round-trip when checking commited content
This restores a previous behaviour of the autosave helper where it
didn't bother 'saving' the content of the editor when that operation
was redundant. Since we no longer have a local cache of the last
autosaved value we still need to send the content in those cases,
but we _can_ avoid the trip when checking content known to be the
same as the current revision.

Change-Id: Ib0d7744492c84aedaaacc759f2d407ac4f328999