Extend tests for fileTreeCompat before changing it
This adds testing of the initial tree as well as empty directories.
Directories in general are moved to a similar mechanism as is used
for their creation by the 'file/mkdir' endpoint.

The initial (simpler) tree occurs both before any content is added
as well as afterwards (the latter using a revision specifier to
also validate the revision handling).

Change-Id: If1984565282cd8ef5efd36387fc97b2d85bfc59d