Rewrite fileTreeCompat to use git only
This means that it can be run on bare repos as well as per-user
ones now that it's not having to load information about autosaves.
It's not clear what performance impact this has for the moment, but
it does avoid getting a write lock for a conceptually read-only
operation and should thus avoid conflict with other operations.

Change-Id: I510ec7e587e87ca38c3728d96f63c660eab67dd9