Explicitly persist directories
Previously empty directories were not be persisted within the IDE,
so now we create a placeholder '.directory' file in the root of
every folder in the repo which tracks its existence.

While this departes from the way Git represents things, it's nicer
for users since:
 - it's more like how ordinary file-systems work
 - it means that empty folders can be shared with other team-members

It also removes the reliance on the file-system as part of tracking
the repository content, enabling another portion of our migration
to a pure-git system.

Change-Id: Id682d8d18f5be31da7e87573411bcff5760ca1e1