Explicitly trust our piwik scripts
The 'self' policy requires an exact match on the origin of items,
so www.studentrobotics.org does not match studentrobotics.org and
vice-versa. As a result, the loading of the piwik scripts from one
of these results in them not working on the 'other' domain.

Since both are valid (and possible to be used; there are no redirects
in place at the /ide/ level, even though the main website now
redirects to studentrobotics.org) we need to ensure that our piwik
script is reliably available. Adding the specific domain it uses
is a simple fix.

Change-Id: I4a262b459b55c288fd4a0e724aa2fc33d72b62cb