Remove more redundant actions when linting a file
This removes interactions with the per-user-clone which we don't
actually end up using anyway. Just create a fresh clone from the
master repo and then checkout the right version from there.

This makes checking the existence of the file a little more complex
but appears to offer about a 20% speedup for tiny files.

Change-Id: Idf6133e6cf8d8b85fdca8e97b19ca653fac9ec57